Added Tree and Testing modules, as well as a Fix for the .gitignore
The Tree module has the Node and Tree classes to be used for inheritance. The Testing module is a custom implementation of Tree.Node intended to test the AlphaBeta algorithm inside the Tree.Tree class. The fix related to the .gitignore file is meant to avoid common temporary files in python